+++ praktikum-4/src/rumus/Geometry.java
@@ -1,8 +1,7 @@
 package rumus;
-import java.util.Scanner;
 
 public abstract class Geometry {
-    protected abstract double volume();
-    protected abstract double surfaceArea();
+    protected abstract int volume();
+    protected abstract int surfaceArea();
 
 }
